Move main tests into main_test.py
